testando para ordenação por inserção

+++ src/BubbleSort.java
@@ -10,12 +10,12 @@ public class BubbleSort<T extends Comparable<T>> implements IOrdenador<T>{
     private double nanoToMilli = 1.0/1_000_000;
 
     @Override
-    public int getComparacoes() {
+    public long getComparacoes() {
         return comparacoes;
     }
 
     @Override
-    public int getMovimentacoes() {
+    public long getMovimentacoes() {
         return movimentacoes;
     }
